requirements.txt
Django==2.2.2
django-extensions==2.1.9
djangorestframework==3.10.2
graphviz==0.11.1
mysql-connector-python==8.0.16
protobuf==3.8.0
pydot==1.4.1
pyparsing==2.4.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
